move OFT functionality to SpokePool for easy further integration and removing boilerplate code
