+++ contracts/Succinct_SpokePool.sol
@@ -8,6 +8,9 @@ import "./external/interfaces/SuccinctInterfaces.sol";
  * @notice Succinct Spoke pool.
  */
 contract Succinct_SpokePool is SpokePool, ITelepathyHandler {
+    // Scroll_SpokePool does not use OFT messaging, setting the cap to 0
+    uint256 private constant OFT_FEE_CAP = 0;
+
     // Address of the succinct AMB contract.
     address public succinctTargetAmb;
 
@@ -47,7 +50,7 @@ contract Succinct_SpokePool is SpokePool, ITelepathyHandler {
         address _wrappedNativeTokenAddress,
         uint32 _depositQuoteTimeBuffer,
         uint32 _fillDeadlineBuffer
-    ) SpokePool(_wrappedNativeTokenAddress, _depositQuoteTimeBuffer, _fillDeadlineBuffer) {} // solhint-disable-line no-empty-blocks
+    ) SpokePool(_wrappedNativeTokenAddress, _depositQuoteTimeBuffer, _fillDeadlineBuffer, OFT_FEE_CAP) {} // solhint-disable-line no-empty-blocks
 
     /**
      * @notice Construct the Succinct SpokePool.
