+++ package.json
@@ -49,9 +49,6 @@
     "@ethersproject/abstract-provider": "5.7.0",
     "@ethersproject/abstract-signer": "5.7.0",
     "@ethersproject/bignumber": "5.7.0",
-    "@layerzerolabs/lz-evm-protocol-v2": "^3.0.70",
-    "@layerzerolabs/oapp-evm": "^0.3.1",
-    "@layerzerolabs/oft-evm": "^3.1.2",
     "@openzeppelin/contracts": "4.9.6",
     "@openzeppelin/contracts-upgradeable": "4.9.6",
     "@scroll-tech/contracts": "^0.1.0",
