+++ storage-layouts/AlephZero_SpokePool.json
@@ -152,9 +152,15 @@
     },
     {
       "contract": "contracts/AlephZero_SpokePool.sol:AlephZero_SpokePool",
-      "label": "__gap",
+      "label": "oftMessengers",
       "offset": 0,
       "slot": "2164"
+    },
+    {
+      "contract": "contracts/AlephZero_SpokePool.sol:AlephZero_SpokePool",
+      "label": "__gap",
+      "offset": 0,
+      "slot": "2165"
     },
     {
       "contract": "contracts/AlephZero_SpokePool.sol:AlephZero_SpokePool",
@@ -170,15 +176,9 @@
     },
     {
       "contract": "contracts/AlephZero_SpokePool.sol:AlephZero_SpokePool",
-      "label": "oftMessengers",
-      "offset": 0,
-      "slot": "3164"
-    },
-    {
-      "contract": "contracts/AlephZero_SpokePool.sol:AlephZero_SpokePool",
       "label": "__gap",
       "offset": 0,
-      "slot": "3165"
+      "slot": "3164"
     }
   ]
 }
