+++ storage-layouts/Optimism_SpokePool.json
@@ -152,9 +152,15 @@
     },
     {
       "contract": "contracts/Optimism_SpokePool.sol:Optimism_SpokePool",
-      "label": "__gap",
+      "label": "oftMessengers",
       "offset": 0,
       "slot": "2164"
+    },
+    {
+      "contract": "contracts/Optimism_SpokePool.sol:Optimism_SpokePool",
+      "label": "__gap",
+      "offset": 0,
+      "slot": "2165"
     },
     {
       "contract": "contracts/Optimism_SpokePool.sol:Optimism_SpokePool",
