+++ storage-layouts/Redstone_SpokePool.json
@@ -152,9 +152,15 @@
     },
     {
       "contract": "contracts/Redstone_SpokePool.sol:Redstone_SpokePool",
-      "label": "__gap",
+      "label": "oftMessengers",
       "offset": 0,
       "slot": "2164"
+    },
+    {
+      "contract": "contracts/Redstone_SpokePool.sol:Redstone_SpokePool",
+      "label": "__gap",
+      "offset": 0,
+      "slot": "2165"
     },
     {
       "contract": "contracts/Redstone_SpokePool.sol:Redstone_SpokePool",
