+++ storage-layouts/Scroll_SpokePool.json
@@ -152,9 +152,15 @@
     },
     {
       "contract": "contracts/Scroll_SpokePool.sol:Scroll_SpokePool",
-      "label": "__gap",
+      "label": "oftMessengers",
       "offset": 0,
       "slot": "2164"
+    },
+    {
+      "contract": "contracts/Scroll_SpokePool.sol:Scroll_SpokePool",
+      "label": "__gap",
+      "offset": 0,
+      "slot": "2165"
     },
     {
       "contract": "contracts/Scroll_SpokePool.sol:Scroll_SpokePool",
