+++ storage-layouts/WorldChain_SpokePool.json
@@ -152,9 +152,15 @@
     },
     {
       "contract": "contracts/WorldChain_SpokePool.sol:WorldChain_SpokePool",
-      "label": "__gap",
+      "label": "oftMessengers",
       "offset": 0,
       "slot": "2164"
+    },
+    {
+      "contract": "contracts/WorldChain_SpokePool.sol:WorldChain_SpokePool",
+      "label": "__gap",
+      "offset": 0,
+      "slot": "2165"
     },
     {
       "contract": "contracts/WorldChain_SpokePool.sol:WorldChain_SpokePool",
